Add checks for valid domain to passive dns lookups.
This fixes a bug where including a scheme in a query, like
'https://www.example.com' would cause the entire url to be sent to
passivetotal. Passivetotal was responding with 500 errors. Now inputs to
the passive dns lookup checks for a valid, parsable domain or ipv4
address before querying the Passivetotal API.